+++ module/Config/src/Config/Form/Config.php
@@ -68,6 +68,7 @@ class Config extends AbstractForm
         $generalFieldset->setLabel('General');
         $name = new Element\Text('site_name');
         $name->setAttribute('label', 'Site name')
+            ->setAttribute('id', 'site_name')
             ->setAttribute('class', 'input-text');
         $generalFieldset->add($name);
 
@@ -102,6 +103,7 @@ class Config extends AbstractForm
         $offlineDocument = new Element\Select('site_offline_document');
         $offlineDocument->setAttribute('label', 'Offline document')
             ->setAttribute('class', 'input-select')
+            ->setAttribute('id', 'site_offline_document')
             ->setValueOptions(array('Select document') + $documentCollection->getSelect());
         $generalFieldset->add($offlineDocument);
 
@@ -117,6 +119,7 @@ class Config extends AbstractForm
         $layoutNotFound   = new Element\Select('site_404_layout');
         $layoutNotFound->setAttribute('label', '404 layout')
             ->setAttribute('class', 'input-select')
+            ->setAttribute('id', 'site_404_layout')
             ->setValueOptions(array('Select document') + $layoutCollection->getSelect());
         $generalFieldset->add($layoutNotFound);
 
@@ -131,6 +134,7 @@ class Config extends AbstractForm
         $layoutException = new Element\Select('site_exception_layout');
         $layoutException->setAttribute('label', 'Exception layout')
             ->setAttribute('class', 'input-select')
+            ->setAttribute('id', 'site_exception_layout')
             ->setValueOptions(array('Select document') + $layoutCollection->getSelect());
         $generalFieldset->add($layoutException);
         $this->getInputFilter()->add(
@@ -158,6 +162,7 @@ class Config extends AbstractForm
         $sessionFieldset->setLabel('Session');
         $cookieDomain = new Element\Text('cookie_domain');
         $cookieDomain->setAttribute('label', 'Cookie domain')
+            ->setAttribute('id', 'cookie_domain')
             ->setAttribute('class', 'input-text');
         $sessionFieldset->add($cookieDomain);
 
@@ -174,6 +179,7 @@ class Config extends AbstractForm
 
         $cookiePath = new Element\Text('cookie_path');
         $cookiePath->setAttribute('label', 'Cookie path')
+            ->setAttribute('id', 'cookie_path')
             ->setAttribute('class', 'input-text');
         $sessionFieldset->add($cookiePath);
 
@@ -191,6 +197,7 @@ class Config extends AbstractForm
         $sessionHandler = new Element\Select('session_handler');
         $sessionHandler->setAttribute('label', 'Session handler')
             ->setAttribute('class', 'input-select')
+            ->setAttribute('id', 'session_handler')
             ->setValueOptions(array('0' => 'Files', '1' => 'Database'));
         $sessionFieldset->add($sessionHandler);
 
@@ -207,6 +214,7 @@ class Config extends AbstractForm
 
         $sessionPath = new Element\Text('session_path');
         $sessionPath->setAttribute('label', 'Session path')
+            ->setAttribute('id', 'session_path')
             ->setAttribute('class', 'input-text');
         $sessionFieldset->add($sessionPath);
 
@@ -220,6 +228,7 @@ class Config extends AbstractForm
 
         $sessionLifetime = new Element\Text('session_lifetime');
         $sessionLifetime->setAttribute('label', 'Session lifetime')
+            ->setAttribute('id', 'session_lifetime')
             ->setAttribute('class', 'input-text');
         $sessionFieldset->add($sessionLifetime);
 
@@ -242,6 +251,7 @@ class Config extends AbstractForm
         $debugFieldset->setLabel('Debug');
         $debugIsActive = new Element\Checkbox('debug_is_active');
         $debugIsActive->setAttribute('label', 'Debug is active')
+            ->setAttribute('id', 'debug_is_active')
             ->setAttribute('id', 'input-checkbox')
             ->setAttribute('class', 'input-checkbox');
         $debugFieldset->add($debugIsActive);
@@ -264,6 +274,7 @@ class Config extends AbstractForm
         $cacheFieldset->setLabel('Cache');
         $cacheIsActive = new Element\Checkbox('cache_is_active');
         $cacheIsActive->setAttribute('label', 'Cache is active')
+            ->setAttribute('id', 'cache_is_active')
             ->setAttribute('class', 'input-checkbox')
             ->setAttribute('id', 'cache-active');
         $cacheFieldset->add($cacheIsActive);
@@ -281,6 +292,7 @@ class Config extends AbstractForm
 
         $cacheHandler = new Element\Select('cache_handler');
         $cacheHandler->setAttribute('class', 'input-select')
+            ->setAttribute('id', 'cache_handler')
             ->setAttribute('label', 'Cache handler');
         $handlerWhitelist = array('filesystem' => 'FileSystem');
         if (extension_loaded('apc')) {
@@ -307,6 +319,7 @@ class Config extends AbstractForm
 
         $cacheLifetime = new Element\Text('cache_lifetime');
         $cacheLifetime->setAttribute('label', 'Cache lifetime')
+            ->setAttribute('id', 'cache_lifetime')
             ->setAttribute('class', 'input-text');
         $cacheFieldset->add($cacheLifetime);
 
@@ -341,6 +354,7 @@ class Config extends AbstractForm
         $localeFieldset->setLabel('Locale');
         $locale = new Element\Select('locale');
         $locale->setAttribute('label', 'Server locale')
+            ->setAttribute('id', 'locale')
             ->setAttribute('class', 'input-select')
             ->setValueOptions($config['locales']);
         $localeFieldset->add($locale);
@@ -363,6 +377,7 @@ class Config extends AbstractForm
         $mailFieldset->setLabel('Mail');
         $mailFrom = new Element\Text('mail_from');
         $mailFrom->setAttribute('label', 'From E-mail')
+            ->setAttribute('id', 'mail_from')
             ->setAttribute('class', 'input-text');
         $mailFieldset->add($mailFrom);
 
@@ -379,6 +394,7 @@ class Config extends AbstractForm
 
         $mailFromName = new Element\Text('mail_from_name');
         $mailFromName->setAttribute('label', 'From name')
+            ->setAttribute('id', 'mail_from_name')
             ->setAttribute('class', 'input-text');
         $mailFieldset->add($mailFromName);
 
@@ -401,6 +417,7 @@ class Config extends AbstractForm
 
         $forceBackendSsl = new Element\Checkbox('force_backend_ssl');
         $forceBackendSsl->setAttribute('label', 'Force backend SSL')
+            ->setAttribute('id', 'force_backend_ssl')
             ->setAttribute('class', 'input-checkbox')
             ->setAttribute('id', 'force-backend-ssl');
         $webFieldset->add($forceBackendSsl);
@@ -415,6 +432,7 @@ class Config extends AbstractForm
 
         $forceFrontendSsl = new Element\Checkbox('force_frontend_ssl');
         $forceFrontendSsl->setAttribute('label', 'Force frontend SSL')
+            ->setAttribute('id', 'force_frontend_ssl')
             ->setAttribute('class', 'input-checkbox')
             ->setAttribute('id', 'force-frontend-ssl');
         $webFieldset->add($forceFrontendSsl);
@@ -439,6 +457,7 @@ class Config extends AbstractForm
         foreach ($pathFields as $label => $identifier) {
             $field = new Element\Text($identifier);
             $field->setAttribute('label', $label)
+                ->setAttribute('id', $identifier)
                 ->setAttribute('class', 'input-text');
             $webFieldset->add($field);